r6tracker::types::user_data

Struct UserData

pub struct UserData {
    pub id: String,
    pub kind: String,
    pub children: Vec<Season>,
    pub metadata: Metadata,
    pub stats: Vec<Stat>,
    pub unknown: HashMap<String, Value>,
}

Fields
id: String

Unique user id

kind: Stringchildren: Vec<Season>

Collection of ranked seasons stats

metadata: Metadata

Metadata

stats: Vec<Stat>

A collection of all stats

unknown: HashMap<String, Value>

Unknown fields

Implementations

impl UserData

pub fn current_mmr(&self) -> Option<u32>

Gets top mmr from all servers.

pub fn current_mmr_image(&self) -> Option<&Url>

Get the image url for the rank this user is at gloablly

pub fn current_mmr_america(&self) -> Option<u32>

Get the MMR for this user.

pub fn season_color(&self) -> &str

Gets this season’s color as a string hex value

pub fn season_color_u32(&self) -> Option<u32>

Tries to parse this season’s hex color as a u32

pub fn get_kills(&self) -> Option<u64>

Get total # of kills

pub fn get_deaths(&self) -> Option<u64>

Get total # of deaths

pub fn kd(&self) -> Option<f64>

Get overall K/D

pub fn wl(&self) -> Option<f64>

Get Overall W/L

pub fn name(&self) -> &str

Get user tag name

pub fn avatar_url(&self) -> &Url

Get user avatar url

pub fn get_latest_season(&self) -> Option<&Season>

Get the latest stats for the latest ranked region/season the user has played in

pub fn get_max_season(&self) -> Option<&Season>

Get the season where the user attained their max ranking